# core/src/main/resources/META-INF/openapi.yml
openapi: 3.0.0
servers:
  - url: http://192.168.4.102:80
info:
  version: 0.3.0
  title: RootCause - API
  description: |2-
     API designed to use Root Cause features.  
     Implemented features:                     
     - [POST] insertlog
     - [GET]  report/{id}
     - [GET]  link
     - [GET]  tokens
     - [GET]  tokentypes
     
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /insertlog:
    post:
      tags:
        - Insertion
      description: Adds a log to the database
      operationId: insertLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertionLogFormat'
      responses:
        '200':
          description: Successful insertion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsertionLogResponse'
        '400':
          description: Invalid parameter. Insertion failed
...
components:
  schemas:
...
    InsertionLogFormat:
      type: array
      items:
        type: object
        required:
          - log
        properties:
          log:
            type: string
    InsertionLogResponse:
      type: array
      items:
        type: object
        required:
          - id
        properties:
          id:
            type: integer
...